Reorganized the rest of test classes into nested test classes
