--- src/test/java/com/ansill/arrays/ByteBufferByteArrayTest.java
@@ -1,9 +1,0 @@
-package com.ansill.arrays;
-
-import org.junit.jupiter.api.DisplayName;
-import test.ByteArrayTest;
-
-@DisplayName("ByteBufferByteArray tests")
-public interface ByteBufferByteArrayTest extends ByteArrayTest{
-
-}
