--- src/test/java/com/ansill/arrays/ReadOnlyByteBufferByteArrayWithControlByteArrayTest.java
@@ -1,11 +1,0 @@
-package com.ansill.arrays;
-
-import org.junit.jupiter.api.DisplayName;
-import test.ReadOnlyByteArrayWithOtherByteArrayTest;
-
-@DisplayName("ByteBufferByteArray - ReadOnly test with control ByteArray implementation")
-public
-class ReadOnlyByteBufferByteArrayWithControlByteArrayTest extends ReadOnlyByteBufferByteArrayTest implements
-  ReadOnlyByteArrayWithOtherByteArrayTest{
-
-}
