--- src/test/java/com/ansill/arrays/ReadOnlyMultipleByteArrayWithControlByteArrayTest.java
@@ -1,11 +1,0 @@
-package com.ansill.arrays;
-
-import org.junit.jupiter.api.DisplayName;
-import test.ReadOnlyByteArrayWithOtherByteArrayTest;
-
-@DisplayName("MultipleByteArray - ReadOnly test with control ByteArray implementation")
-public
-class ReadOnlyMultipleByteArrayWithControlByteArrayTest extends ReadOnlyMultipleByteArrayTest implements
-  ReadOnlyByteArrayWithOtherByteArrayTest{
-
-}
