--- src/test/java/com/ansill/arrays/WriteOnlyByteBufferByteArrayWithByteBufferByteArrayTest.java
@@ -1,17 +1,0 @@
-package com.ansill.arrays;
-
-import org.junit.jupiter.api.DisplayName;
-
-import javax.annotation.Nonnull;
-
-@DisplayName("ByteBufferByteArray - WriteOnly test with PrimitiveByteArray implementation")
-public
-class WriteOnlyByteBufferByteArrayWithByteBufferByteArrayTest
-  extends WriteOnlyByteBufferByteArrayWithControlByteArrayTest{
-
-  @Nonnull
-  @Override
-  public ReadableWritableByteArray createControlReadableWritable(long size){
-    return new PrimitiveByteArray(new byte[(int) size]);
-  }
-}
