--- src/test/java/com/ansill/arrays/WriteOnlyMultipleByteArrayWithControlByteArrayTest.java
@@ -1,11 +1,0 @@
-package com.ansill.arrays;
-
-import org.junit.jupiter.api.DisplayName;
-import test.WriteOnlyByteArrayWithOtherByteArrayTest;
-
-@DisplayName("MultipleByteArray - WriteOnly test with control ByteArray implementation")
-public
-class WriteOnlyMultipleByteArrayWithControlByteArrayTest extends WriteOnlyMultipleByteArrayTest implements
-  WriteOnlyByteArrayWithOtherByteArrayTest{
-
-}
